AlphaHeavy.Time

Description

4 corner simultaneous 24 hour Days that occur within a single 4 corner rotation of Earth.

Documentation

class DateTime f where

Associated Types

type DateTimeZone f :: TimeZone

The timezone associated with this type

data DateTimeComponents f :: *

The natural components of the time. For a regular datetime this may represent days/minutes/hours/etc. For durations it will be in seconds or an equivalent unit

Methods

unpack :: f -> DateTimeComponents f

Unpack a time into its components

pack :: DateTimeComponents f -> f

Repack a time from its components

newtype DateTimeLensT m c f a

Constructors

DateTimeLensT

Fields

unDateTimeLens :: Lens (Kleisli (StateT c m)) f a

data TimeZone

Constructors

UTC

LocalTime

newtype TimeZoneOffset

Constructors

TimeZoneOffset

Fields

getTimeZoneOffset :: Int32

Date storage

newtype UnixTime t

Constructors

UnixTime Int64

newtype UnixTimeNanos t

Constructors

UnixTimeNanos (Int64, Int32)

data JavaTime t

data JulianDate t

System time

getCurrentUnixTime :: IO (UnixTime UTC)

getCurrentUnixTimeNanos :: IO (UnixTimeNanos UTC)

Raw date components

data DateTimeStruct

Constructors

DateTimeStruct

Fields

dt_year :: Year

dt_month :: Month

dt_day :: Day

dt_hour :: Hour

dt_minute :: Minute

dt_second :: Second

dt_nanos :: Nano

class DateTimePart c f a where

Functions used to build a DateTimeLensT

Methods

dtg :: (Functor m, Monad m) => f -> StateT c m a

dts :: (Functor m, Monad m) => a -> f -> StateT c m f

datePart :: (Functor m, Monad m, DateTimePart c f a) => DateTimeLensT m c f a

A lens from a time into a time component

Lens accessors

Pure

get :: DateTime f => DateTimeLensT Identity (DateTimeComponents f) f a -> f -> a

set :: DateTime f => DateTimeLensT Identity (DateTimeComponents f) f a -> a -> f -> f

modify :: DateTime f => DateTimeLensT Identity (DateTimeComponents f) f a -> (a -> a) -> f -> f

Monadic (currently not very monadic...)

getM :: (Functor m, Monad m, DateTime f) => DateTimeLensT m (DateTimeComponents f) f a -> f -> m a

getM2 :: (Functor m, Monad m, DateTime f) => (DateTimeLensT m (DateTimeComponents f) f a, DateTimeLensT m (DateTimeComponents f) f b) -> f -> m (a, b)

setM :: (Functor m, Monad m, DateTime f) => DateTimeLensT m (DateTimeComponents f) f a -> a -> f -> m f

modifyM :: (Functor m, Monad m, DateTime f) => DateTimeLensT m (DateTimeComponents f) f a -> (a -> a) -> f -> m f

Helper lenses

year :: (Functor m, Monad m, DateTimePart c f Year) => DateTimeLensT m c f Year

month :: (Functor m, Monad m, DateTimePart c f Month) => DateTimeLensT m c f Month

monthOfYear :: (Functor m, Monad m, DateTimePart c f MonthOfYear) => DateTimeLensT m c f MonthOfYear

week :: (Functor m, Monad m, DateTimePart c f Week) => DateTimeLensT m c f Week

day :: (Functor m, Monad m, DateTimePart c f Day) => DateTimeLensT m c f Day

dayOfWeek :: (Functor m, Monad m, DateTimePart c f DayOfWeek) => DateTimeLensT m c f DayOfWeek

hour :: (Functor m, Monad m, DateTimePart c f Hour) => DateTimeLensT m c f Hour

minute :: (Functor m, Monad m, DateTimePart c f Minute) => DateTimeLensT m c f Minute

second :: (Functor m, Monad m, DateTimePart c f Second) => DateTimeLensT m c f Second

milli :: (Functor m, Monad m, DateTimePart c f Milli) => DateTimeLensT m c f Milli

nano :: (Functor m, Monad m, DateTimePart c f Nano) => DateTimeLensT m c f Nano

pico :: (Monad m, Functor m, DateTimePart c f Pico) => DateTimeLensT m c f Pico

timeZoneOffset :: (Functor m, Monad m, DateTimePart c f TimeZoneOffset) => DateTimeLensT m c f TimeZoneOffset

Date sections

newtype Year

Constructors

Years

Fields

getYear :: Int32

newtype Month

Constructors

Months

Fields

getMonth :: Int32

data MonthOfYear

Constructors

January

February

March

April

May

June

July

August

September

October

November

December

newtype Week

Constructors

Weeks

Fields

getWeek :: Int32

newtype Day

Constructors

Day

Fields

getDay :: Int32

data DayOfWeek

Constructors

Sunday

Monday

Tuesday

Wednesday

Thursday

Friday

Saturday

newtype Hour

Constructors

Hours

Fields

getHour :: Int32

newtype Minute

Constructors

Minutes

Fields

getMinute :: Int32

newtype Second

Constructors

Seconds

Fields

getSecond :: Int32

newtype Milli

Constructors

Millis

Fields

getMillis :: Int64

newtype Nano

Constructors

Nanos

Fields

getNanos :: Int64

newtype Pico

Constructors

Picos

Fields

getPicos :: Int64

Durations

class Duration a